Moved all product and cart typings to appropiate modules
