+++ core/modules/cart/components/Microcart.ts
@@ -1,6 +1,6 @@
-import AppliedCoupon from '@vue-storefront/store/types/cart/AppliedCoupon'
-import Product from '@vue-storefront/store/types/product/Product'
-import CartTotalSegments from '@vue-storefront/store/types/cart/CartTotalSegments'
+import AppliedCoupon from '../types/AppliedCoupon'
+import Product from '@vue-storefront/core/modules/catalog/types/Product'
+import CartTotalSegments from '../types/CartTotalSegments'
 
 export const Microcart = {
   name: 'Microcart',
