+++ core/modules/cart/types/CartItem.ts
@@ -1,4 +1,4 @@
-import Product from '@vue-storefront/store/types/product/Product'
+import Product from '@vue-storefront/core/modules/catalog/types/Product'
 
 import CartItemOption from './CartItemOption'
 import CartItemTotals from './CartItemTotals'
