+++ core/modules/cart/types/CartTotalSegmentsItem.ts
