+++ core/modules/catalog/types/Product.ts
