+++ doc/TypeScript Action Plan.md
@@ -25,8 +25,8 @@ We're in the middle of [refactoring `core/components` to `core/api` modules](htt
 1. Introduce types - move *.js modules to *.ts modules incrementally without breaking changes. 
 2. Use Types when it's approprierate in Your newly written modules and new features.
 3. One Vuex module, or just few components refactored within one release (once a month) is fine.
-4. All `core/api` modules should be created using TypeScript.
+4. All Vue Storefront modules should be created using TypeScript.
 5. All new modules and vuex stores should be created using TypeScript.
 6. **For now please don't refactor existing UI layer (components, pages) to use TypeScript. We should focus at Vuex, core libraries and API's at first to not introduce a chaos into theme development.**
-6. We should put the types/interfaces inside `core/store/types` for all Entity/Data related models or in `core/types` for some shared one's; for example: Product.ts should be placed in `core/store/types/product/Product.ts`
+6. We should put the types/interfaces inside appropiate modules inside `types` for all Entity/Data related models.
 7. We should use minimal possible set of interfaces. Try to introduce 1-2 interfaces per entity (e.g. Product shouldn't be represented by more than 2 interfaces)
