+++ src/themes/default/components/core/AddToCart.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script lang="ts">
-import Product from '@vue-storefront/store/types/product/Product'
+import Product from '@vue-storefront/core/modules/catalog/types/Product'
 import { formatProductMessages } from '@vue-storefront/core/filters/product-messages/typed'
 import focusClean from 'theme/components/theme/directives/focusClean'
 import ButtonFull from 'theme/components/theme/ButtonFull.vue'
